requirements.txt
jieba
msgpack-rpc-python
six
